fix nesting button
