+++ packages/react-provider/src/reducer.ts
@@ -13,6 +13,7 @@ export const registerReducer = (scope, reducer) => {
 };
 
 const rootReducer = (state, action) => {
+  console.log(action);
   if (
     typeof window !== "undefined" &&
     (
